fix: Resolve merge conflict

+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/navigation/MypageNavigation.kt
@@ -18,7 +18,7 @@ fun NavController.navigateToGainedCharacter() {
 }
 
 fun NavGraphBuilder.myPageNavGraph(
-    navigateToGainedCharacter: (Int) -> Unit,
+    navigateToGainedCharacter: () -> Unit,
     navigateToBack: () -> Unit,
 ) {
     composable<MainTabRoute.MyPage> {
